requirements.txt
# Discord Bot Requirements
discord.py>=2.3.0
python-dotenv>=1.0.0

# Music functionality
yt-dlp>=2023.7.6

# Additional utilities
aiohttp>=3.8.0
psutil>=5.9.0

# Optional: For voice encryption
PyNaCl>=1.5.0

# News functionality
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Future features
requests>=2.31.0

# AI functionality - DeepSeek
openai>=1.80.0
